+++ CHANGELOG.md
@@ -39,6 +39,11 @@ is based on [Keep a Changelog](https://keepachangelog.com).
   `--foo=true` is now equivalent to `--foo` and `--foo=false` will set the flag
   to `false`. Short options like `-f` only accept the argument when passing it
   without a space, e.g., `-ffalse`.
+- The `buffer` and `interval` operators now properly check the time period
+  parameter and produce an error if the period is zero or negative (#2030).
+  Passing an invalid delay parameter to `to_stream` or `to_typed_stream`
+  likewise produces a stream that immediately calls `on_error` on any client
+  that tries to subscribe to it.
 
 ## [1.0.2] - 2024-10-30